refactor: move the top-level bench result cluster to homeboy-extension-contract
The finale of the bench subsystem breakdown. With every field dependency now in
a contract or leaf crate (via the preceding cuts + the observation::timeline
relocation), the four top-level bench result types move wholesale:
BenchResults, BenchRunMetadata, BenchScenario, BenchRunSnapshot.

BenchResults is THE most-imported bench symbol across core. Moving it - and the
whole result_types graph it anchors - collapses the largest reverse-edge surface
of the extension subsystem. The pure custom deserializer deserialize_budget_findings
(just Vec<HomeboyFinding> deserialization) moves with the types; the budget-finding
construction helpers (failure, is_gate_failure) stay in core.

Added homeboy-finding as a direct contract dep. Core's result_types.rs is now a
thin re-export of the full bench result type graph from the contract crate, so
all ~33 consumers and the parsing:: re-export surface stay stable.
